# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API Despesas Cartão",
        "description": "Descrição da API para gerenciar despesas de cartão",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cards_users/": {
            "post": {
                "tags": [
                    "Cartões"
                ],
                "summary": "Insere na base de dados os registros do novo cartão.",
                "operationId": "72fd541d1a7867da985024d0cb9b9378",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    },
                    "425": {
                        "description": "Undefined User!"
                    },
                    "428": {
                        "description": "Card already registered for the user!"
                    }
                }
            }
        },
        "/api/cards_users/{id_user}": {
            "get": {
                "tags": [
                    "Cartões"
                ],
                "summary": "Lista todos os cartões, somente do usuário selecionado.",
                "operationId": "3900c1b1c0c1816be152f63e1e63bd69",
                "responses": {
                    "420": {
                        "description": "No registration!"
                    },
                    "425": {
                        "description": "Undefined User!"
                    }
                }
            }
        },
        "/api/cards_users/{id_card}": {
            "put": {
                "tags": [
                    "Cartões"
                ],
                "summary": "Atualiza os dados somente do cartão selecionado já cadastrado.",
                "operationId": "9ad2208248c8b986d2245318ca177317",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    },
                    "428": {
                        "description": "Card already registered for the user!"
                    },
                    "429": {
                        "description": "Undefined Card!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cartões"
                ],
                "summary": "Exclui na base de dados os registros do cartão selecionado.",
                "operationId": "ef5bb2d0292a91fd4e8fb604a19f2a01",
                "responses": {
                    "200": {
                        "description": "Successfully Done!"
                    },
                    "500": {
                        "description": "Unexpected error!"
                    },
                    "424": {
                        "description": "Request error, validate the data and try again!"
                    },
                    "429": {
                        "description": "Undefined Card!"
                    }
                }
            }
        },
...
    },
...
}